*  Expose all I/O modes in SIS 3820 *  Fix some  overly large read buffers all over the place. *  Centralize firmware loads for FERA XLM and base class. *  In bus ownership register only look at the bottom 2bits    since it's a 2 bit register.

+++ main/usb/common/devices/CULMTrigger.hpp
@@ -295,7 +295,7 @@ CULMTrigger<Controller,RdoList>::loadFirmware2(Controller& controller,
 
     // load body of firmware into rdolist
     RdoList* body = controller.createReadoutList();
-    uint32_t dummy;
+    uint16_t dummy;
     size_t nbytes;
     int status = -100;
